Add tests for cycle tags & Refactoring
* Add unit tests for cycling the tags
* Move all filter tests to their own module
* Remove panic if tags in filter aren't in current tags because it's an
  unwritten contract
* Update readme
* Remove not needed derive
